Moved the initialization of the crashlytics logger to the logger itself
