+++ app/src/main/java/com/xmartlabs/template/module/AndroidModule.java
@@ -3,7 +3,7 @@ package com.xmartlabs.template.module;
 import android.app.Application;
 
 import com.xmartlabs.base.core.module.CoreAndroidModule;
-import com.xmartlabs.template.BuildInfo;
+import com.xmartlabs.template.model.BuildInfo;
 
 public class AndroidModule extends CoreAndroidModule {
   public AndroidModule(Application application) {
